s32: s32z: patch rtd 2.0.1 to work with zephyr
Remove 'u' suffix from macros so that they can be used with
LISTIFY in shim drivers.

Remove 'static inline' keywords from 'Netc_Eth_Ip_MSIX_Rx'
function so that it can be used as an extern function outside
of the file declaration.

Clear the EOC flag before Adc_Sar_CheckAndCallEocNotification()
is called to avoid the race condition that prevents the next end
of conversion callback to be executed, when a conversion is
started inside the previous end of conversion completed callback.

Signed-off-by: Tu Nguyen Van <tu.nguyenvan@nxp.com>