s32: drivers: s32k3: mcl: fix emios mcl ip init
Due to using VAR_SEC_NOCACHE before declaring Emios_Ip_axIpIsInitialized
when EMIOS_IP_MULTICORE_IS_AVAILABLE is not STD_ON, the
Emios_Ip_axIpIsInitialized struct can't be initialized statically. To
initialize Emios_Ip_axIpIsInitialized only one time when the init
function is called, a static boolean flag is used such that initializing
the values within Emios_Ip_axIpIsInitialized occurs once. Further during
the initialization of Emios_Ip_axIpIsInitialized, only the
instanceInitState is set to an initial value, as runCore is not in the
structure for this specific condition.

Signed-of-by: Tori Koelewyn <tori.koelewyn@gmail.com>